// apps/server/src/generated/prisma/models/AiUserHunterConfig.ts
export type AiUserHunterConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$AiUserHunterConfigPayload>
// ...
export type $AiUserHunterConfigPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "AiUserHunterConfig"
  objects: {
    user: Prisma.$SystemUserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    title: string
    apiBase: string
    apiKey: string
    encryptedApiKey: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["aiUserHunterConfig"]>
  composites: {}
}